+++ src/core/animate.ts
@@ -237,6 +237,11 @@ function animateWithCssTransition({
   registerPropertyIfNeeded()
 
   const style = mapValues(fromTo, ([from, to], key) => {
+    // Skip animation if the value is not consistent
+    if (from.values.length !== to.values.length) {
+      return interpolateParsedStyle(to, to.values)
+    }
+
     const style = zip(from.values, to.values).map(([from, to], i) => {
       const initialVelocity = velocity?.[key]?.[i] ?? 0
       return springStyle(spring, {
@@ -282,6 +287,11 @@ function animateWithRaf({
     }
 
     const style = mapValues(fromTo, ([from, to], key) => {
+      // Skip animation if the value is not consistent
+      if (from.values.length !== to.values.length) {
+        return interpolateParsedStyle(to, to.values)
+      }
+
       const realValue = context.realValue[key]
       if (!realValue) {
         return ''
@@ -299,7 +309,7 @@ function animateWithRaf({
     requestAnimationFrame(render)
   }
 
-  requestAnimationFrame(render)
+  render()
 }
 
 function createContext<