+++ src/vue/SpringTransitionGroup.ts
@@ -2,6 +2,7 @@ import {
   ComponentOptions,
   PropType,
   TransitionGroup,
+  TransitionGroupProps,
   VNode,
   computed,
   defineComponent,
@@ -54,7 +55,6 @@ interface Position {
 
 const positionMap = new WeakMap<VNode, Position>()
 const newPositionMap = new WeakMap<VNode, Position>()
-const velocityMap = new WeakMap<VNode, Record<string, number[]>>()
 
 const SpringTransitionGroup = defineComponent({
   name: 'SpringTransitionGroup',
@@ -63,6 +63,7 @@ const SpringTransitionGroup = defineComponent({
 
   setup(props, ctx) {
     const { emit } = ctx
+    const attrs: TransitionGroupProps = ctx.attrs
 
     let prevChildren: VNode[] = []
     let children: VNode[] = []
@@ -80,8 +81,8 @@ const SpringTransitionGroup = defineComponent({
 
     const baseRender = (TransitionGroup as unknown as ComponentOptions).setup!(
       {
+        ...attrs,
         tag: props.tag,
-        css: false,
         onBeforeEnter,
         onEnter,
         onAfterEnter,
@@ -128,16 +129,22 @@ const SpringTransitionGroup = defineComponent({
         const el = child.el as HTMLElementWithController
         const controller = el[scKey]
         if (controller) {
-          // Save current velocity to apply it to the next animation.
-          velocityMap.set(child, { ...controller.realVelocity })
-
-          // Do not call setStyle here as it records wrong velocity.
-          controller.stop()
-
-          createStyleSetter(el)({
-            ...props.springStyle,
-            transform: '',
+          controller.stop({
+            keepVelocity: true,
           })
+
+          controller.setStyle(
+            {
+              ...props.springStyle,
+              transform: '',
+            },
+            {
+              animate: false,
+            },
+          )
+
+          // Clean up enter classes that <TransitionGroup> set.
+          cleanUpEnterClass(el)
         }
       })
 
@@ -183,19 +190,23 @@ const SpringTransitionGroup = defineComponent({
         const el = child.el as HTMLElementWithController
         const controller = el[scKey]!
 
-        const velocity = velocityMap.get(child)
-
-        controller.setStyle(
-          {
-            ...props.springStyle,
-            transform: 'translate(0px, 0px)',
-          },
-          {
-            velocity,
-          },
-        )
+        controller.setStyle({
+          ...props.springStyle,
+          transform: 'translate(0px, 0px)',
+        })
       })
     })
+
+    function cleanUpEnterClass(el: Element): void {
+      const name = attrs.name ?? 'v'
+      const enterFromClass = attrs.enterFromClass ?? `${name}-enter-from`
+      const enterActiveClass = attrs.enterActiveClass ?? `${name}-enter-active`
+      const enterToClass = attrs.enterToClass ?? `${name}-enter-to`
+
+      el.classList.remove(enterFromClass)
+      el.classList.remove(enterActiveClass)
+      el.classList.remove(enterToClass)
+    }
 
     return () => {
       prevChildren = children