+++ test/core/controller.test.ts
@@ -49,6 +49,19 @@ describe('AnimationController', () => {
     expect(actual?.[t]).not.toBe('')
   })
 
+  test('set style without animation if the value is not animatable', () => {
+    let actual: Record<string, string> | undefined
+    const controller = createAnimateController((style) => {
+      actual = style
+    })
+    controller.setOptions({ duration: 10 })
+    controller.setStyle({ width: `auto` })
+    expect(actual?.width).toBe('auto')
+
+    controller.setStyle({ width: `100px` })
+    expect(actual?.width).toBe('100px')
+  })
+
   test('do not trigger animation by setting the same style value', () => {
     let actual: Record<string, string> | undefined
     const controller = createAnimateController((style) => {
@@ -174,6 +187,15 @@ describe('AnimationController', () => {
     controller.setStyle({ transform: `200px 300px` })
     controller.stop()
     expect(controller.realVelocity.transform).toEqual([0, 0])
+  })
+
+  test('realVelocity is kept when stopped with keepVelocity option', () => {
+    const controller = createAnimateController(() => {})
+    controller.setOptions({ duration: 100 })
+    controller.setStyle({ transform: `100px 200px` })
+    controller.setStyle({ transform: `200px 300px` })
+    controller.stop({ keepVelocity: true })
+    expect(controller.realVelocity.transform).not.toEqual([0, 0])
   })
 
   test('realVelocity is calculated from value history without animation', () => {